villian_core: fix undefined host on NON-tty terms
Sometimes the client wont give the username and hostname, especially if the terminal is not a tty. so villian will fail if it tries to get username from url_split.

This commit solves that by having a try block. if we succeed in retrieving hostname and username. then fine, else we set the default value to host, user respectively

Signed-off-by: abuyusif01 <abubakaryusuf951@gmail.com>